+++ Source/IO/WinCopies.IO.Shared/Process/ObjectModel/IProcess.cs
@@ -15,8 +15,6 @@
  * You should have received a copy of the GNU General Public License
  * along with the WinCopies Framework.  If not, see <https://www.gnu.org/licenses/>. */
 
-using WinCopies.Collections.DotNetFix;
-
 namespace WinCopies.IO.Process.ObjectModel
 {
     public interface IProcess : IPropertyObservable